lexidesk-chatbot/requirements.txt
fastapi
uvicorn[standard]
pydantic
sentence-transformers
faiss-cpu
pymupdf
transformers
torch
streamlit
openai
joblib
scikit-learn
sklearn-crfsuite
